+++ src/smallestai/cli/main.py
@@ -6,6 +6,7 @@ from rich.console import Console
 from smallestai.cli.agent_crew import initialise_agent_crew_app
 from smallestai.cli.agents import initialise_agents_app
 from smallestai.cli.auth import initialise_auth_app
+from smallestai.cli.calls import initialise_calls_app
 from smallestai.cli.lib.atoms import AtomsAPIClient
 from smallestai.cli.lib.auth import AuthClient
 from smallestai.cli.lib.project_config import ProjectConfig
@@ -26,6 +27,9 @@ app.add_typer(auth_app, name="auth")
 
 agents_app = initialise_agents_app(auth_client)
 app.add_typer(agents_app, name="agents")
+
+calls_app = initialise_calls_app(auth_client)
+app.add_typer(calls_app, name="calls")
 
 
 def main():